processing/faker/requirements.txt
requests==2.28.1
ruamel.yaml==0.17